changes for project names. rename of folders for new project structure
