+++ core/build.gradle
@@ -5,5 +5,5 @@ sourceCompatibility = 1.6
 sourceSets.main.java.srcDirs = [ "src/" ]
 
 eclipse.project {
-    name = "core"
+    name = projectGroup + "-core"
 }
