+++ core/src/ashley/signals/Signal.java
