+++ tests/assets/coin.png
