+++ tests/assets/crate.png
