+++ tests/build.gradle
@@ -5,5 +5,5 @@ sourceCompatibility = 1.7
 sourceSets.main.java.srcDirs = [ "src/" ]
 
 eclipse.project {
-    name = "ashley-tests"
+    name =  projectGroup + "-tests"
 }
