+++ tests/libs/gdx-backend-lwjgl-natives.jar
