+++ tests/libs/gdx-backend-lwjgl.jar
