+++ tests/src/ashley/tests/components/CameraComponent.java
