+++ tests/src/ashley/tests/components/VisualComponent.java
